// User_app/src/app/interfaces/register.ts
export interface register {
  email: string;
  first_name: string;
  last_name: string;
  password: string;
  gender: string;
  mobile: string;
  fcm_token: string;
  type: string;
  cover: string;
  status: string;
  stripe_key: string;
  country_code: string;
  check: boolean;
  referral: string;
}